01/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
pytest
httpx
requests
